+++ inc/core/core_loader.php
@@ -43,6 +43,7 @@ class Core_Loader {
 			'neve_filter_main_modules',
 			array(
 				'Customizer\Loader',
+				'Views\Tweaks',
 				'Views\Typography',
 				'Views\Top_Bar',
 				'Views\Header',
@@ -72,7 +73,6 @@ class Core_Loader {
 				'Compatibility\Beaver',
 
 				'Admin\Metabox\Manager',
-				'Admin\Metabox\Main',
 			)
 		);
 	}