+++ inc/customizer/options/layout_single_post.php
@@ -50,24 +50,32 @@ class Layout_Single_Post extends Base_Customizer {
 	 * Add content order control.
 	 */
 	private function control_content_order() {
-		$order_default_components = array(
-			'title-meta',
-			'thumbnail',
-			'content',
-			'tags',
+		$order_default_components = apply_filters(
+			'neve_single_post_elements_default_order',
+			array(
+				'title-meta',
+				'thumbnail',
+				'content',
+				'tags',
+				'comments',
+			)
 		);
 
-		$components = array(
-			'title-meta'      => __( 'Title & Meta', 'neve' ),
-			'thumbnail'       => __( 'Thumbnail', 'neve' ),
-			'content'         => __( 'Content', 'neve' ),
-			'tags'            => __( 'Tags', 'neve' ),
-			'post-navigation' => __( 'Post navigation', 'neve' ),
+		$components = apply_filters(
+			'neve_single_post_elements',
+			array(
+				'title-meta'      => __( 'Title & Meta', 'neve' ),
+				'thumbnail'       => __( 'Thumbnail', 'neve' ),
+				'content'         => __( 'Content', 'neve' ),
+				'tags'            => __( 'Tags', 'neve' ),
+				'post-navigation' => __( 'Post navigation', 'neve' ),
+				'comments'        => __( 'Comments', 'neve' ),
+			)
 		);
 
 		$this->add_control(
 			new Control(
-				'neve_single_post_elements_order',
+				'neve_layout_single_post_elements_order',
 				array(
 					'sanitize_callback' => array( $this, 'sanitize_post_elements_ordering' ),
 					'default'           => json_encode( $order_default_components ),
@@ -94,6 +102,9 @@ class Layout_Single_Post extends Base_Customizer {
 			'content',
 			'tags',
 			'post-navigation',
+			'comments',
+			'author-biography',
+			'related-posts',
 		);
 
 		if ( empty( $value ) ) {