+++ inc/views/base_view.php
@@ -20,4 +20,38 @@ abstract class Base_View {
 	 * @return void
 	 */
 	abstract public function init();
+
+	/**
+	 * Get view path to include.
+	 *
+	 * @param string $view_slug the view to be loaded from `views` folder, without extension.
+	 * @param array  $vars      variables used in template.
+	 *
+	 * @return void
+	 */
+	public function get_view( $view_slug, $vars ) {
+		if ( empty( $view_slug ) ) {
+			return;
+		}
+
+		$args = apply_filters( 'neve_filter_view_data_' . $view_slug, $vars );
+
+		$rest_of_path = 'views/' . $view_slug . '.php';
+
+		$path = trailingslashit( get_stylesheet_directory() ) . $rest_of_path;
+
+		if ( file_exists( $path ) && is_readable( $path ) ) {
+			include $path;
+
+			return;
+		}
+
+		$path = trailingslashit( get_template_directory() ) . $rest_of_path;
+
+		if ( file_exists( $path ) && is_readable( $path ) ) {
+			include $path;
+
+			return;
+		}
+	}
 }